Optimize rendering time of plots by making sure snap points do not trigger re-rendering
